Remove claim that downgrade protection isn't needed
While it's true that zone owners may have various motivations for using
several algorithms, the problem with detecting downgrades is unrelated
to the zone owner's intention. That claim is thus besides the point, and
the following doesn't seem right:
> a zone owner can avoid the threat of algorithm downgrade by simply choosing to use only a set of sufficiently strong algorithms

Actually, the problem are cases that turn out to not be in the hand of
the zone owner, because the zone owner knows (as little as the signer)
which algorithms a resolver will support. Let me outline a case where
things are bound to fail (= loss of validation and resolution) when only
the signing rules are relaxed.

Consider transferring a zone from an algorithm 7 DNS provider to an
algorithm 13 provider using a multi-signer setup. During the transition,
the response may contain only signatures of algorithm 7. If used alone,
resolvers not supporting it (such as on Red Hat systems) will treat the
zone as insecure. However, as 13 is also in the DS RRset, such resolvers
will not find a valid path, leading to SERVFAIL.

This issue is independent of the zone owners motivations; it is caused
by the fact that the signer or zone owner can't know what the validator
supports, in combination with the fact that lack of signature(s)
currently only allows insecure treatment when ALL advertised algorithms
are unsupported. Thus, what's needed is not only a change in signing
rules; that change needs to be accompanied by a change in validation
rules to accommodate this case and prevent SERVFAIL.